rename tests package
