+++ final_project/src/test/java/ru/yandex/praktikum/pageObject/HomePageTest.java
@@ -1,4 +1,4 @@
-package ru.yandex.praktikum.pajeObject;
+package ru.yandex.praktikum.pageObject;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.After;
@@ -9,7 +9,6 @@ import org.junit.runners.Parameterized;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import ru.yandex.praktikum.pageObject.HomePage;
 
 import static org.junit.Assert.assertEquals;
 import static ru.yandex.praktikum.pageObject.constants.HomePageConstants.*;
