+++ final_project/src/test/java/ru/yandex/praktikum/pageObject/OrderStatusTest.java
@@ -1,4 +1,4 @@
-package ru.yandex.praktikum.pajeObject;
+package ru.yandex.praktikum.pageObject;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.After;
@@ -6,8 +6,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
-import ru.yandex.praktikum.pageObject.HomePage;
-import ru.yandex.praktikum.pageObject.OrderStatus;
 
 public class OrderStatusTest {
     WebDriver driver;
